# [ready] perf: constant in while in for in busy func