This is a surprisingly big one. Constant definition in while loop in for loop in busy function.

There are other place in the codebase that do something similar (define a const dict for a lookup every time anew).

Also uses `is` instead of `==` for singletons.

There is more to optimize here (use of tuples, doesn't need a `while`) but that's for another MR to keep this small.

```
Before

Total time: 1.51153 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/lazy.py
Function: _push_movement_ops at line 306

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   306                                           @profile
   307                                           def _push_movement_ops(srcs:Tuple[LazyBuffer, ...]) -> Tuple[LazyBuffer, ...]:
   308     80483      22235.8      0.3      1.5    new_srcs = []
   309    144813      40322.9      0.3      2.7    for x in srcs:
   310    144813      29607.2      0.2      2.0      mops: List[Tuple[MovementOps, Any]] = []
   311    144813      22430.4      0.2      1.5      bx = x
   312                                               # backwalk all the movement ops. don't push PAD or EXPAND
   313    144813      94432.1      0.7      6.2      while not bx.realized and bx.optype == MovementOps and bx.op.op != MovementOps.EXPAND and (SHUFFLE_PAD_OPS or bx.op.op != MovementOps.PAD) and len(bx.children) <= 1:
   314     71302      39646.5      0.6      2.6        assert isinstance(bx.op.op, MovementOps)
   315     71302      47581.9      0.7      3.1        mops.append((bx.op.op, bx.op.arg))
   316     71302      59826.3      0.8      4.0        bx = cast(LazyBuffer, bx.op.src[0])
   317                                               # NOTE: can't push pads past anything where f(0, 0) != 0 or f(0) != 0
   318    144813     545717.9      3.8     36.1      unsafe_pad_ops = {BinaryOps.DIV, BinaryOps.CMPLT, UnaryOps.LOG2, UnaryOps.EXP2, UnaryOps.RECIP}
   319    140049      33643.6      0.2      2.2      if mops and not bx.realized and bx.optype == BinaryOps and len(bx.children) <= 1 and (all(x[0] != MovementOps.PAD for x in mops) or all(x.op not in unsafe_pad_ops for x in bx.op.get_lazyops())):
   320      4764     477063.6    100.1     31.6        new_srcs.append(bx.op.replace_with_movement_ops(mops[::-1]))
   321                                               else:
   322    140049      69685.7      0.5      4.6        new_srcs.append(x)
   323     80483      29334.0      0.4      1.9    return tuple(new_srcs)




After

Total time: 0.969821 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/lazy.py
Function: _push_movement_ops at line 308

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   308                                           @profile
   309                                           def _push_movement_ops(srcs:Tuple[LazyBuffer, ...]) -> Tuple[LazyBuffer, ...]:
   310     80483      23872.4      0.3      2.5    new_srcs = []
   311    144813      38574.1      0.3      4.0    for x in srcs:
   312    144813      27994.1      0.2      2.9      mops: List[Tuple[MovementOps, Any]] = []
   313    144813      20276.5      0.1      2.1      bx = x
   314                                               # backwalk all the movement ops. don't push PAD or EXPAND
   315    144813      89660.5      0.6      9.2      while not bx.realized and bx.optype is MovementOps and bx.op.op is not MovementOps.EXPAND and (SHUFFLE_PAD_OPS or bx.op.op is not MovementOps.PAD) and len(bx.children) <= 1:
   316     71302      38714.8      0.5      4.0        assert isinstance(bx.op.op, MovementOps)
   317     71302      46368.5      0.7      4.8        mops.append((bx.op.op, bx.op.arg))
   318     71302      58354.1      0.8      6.0        bx = cast(LazyBuffer, bx.op.src[0])
   319                                               # NOTE: can't push pads past anything where f(0, 0) != 0 or f(0) != 0
   320    140049      30260.9      0.2      3.1      if mops and not bx.realized and bx.optype is BinaryOps and len(bx.children) <= 1 and (all(x[0] is not MovementOps.PAD for x in mops) or all(x.op not in UNSAFE_PAD_OPS for x in bx.op.get_lazyops())):
   321      4764     498316.2    104.6     51.4        new_srcs.append(bx.op.replace_with_movement_ops(mops[::-1]))
   322                                               else:
   323    140049      70097.0      0.5      7.2        new_srcs.append(x)
   324     80483      27332.3      0.3      2.8    return tuple(new_srcs)
```