@tinyb0t 2023-08-27
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/lazy.py                           4               3
------------------------------------------------------------
lines added in the tinygrad folder: 1
```